Moved log1pexp() and randomize_parameters() from core.h into utilities.h
