+++ dlib/dnn/core.h
@@ -114,23 +114,6 @@ namespace dlib
 
 // ----------------------------------------------------------------------------------------
 
-    inline double log1pexp(double x)
-    {
-        using std::exp;
-        using namespace std; // Do this instead of using std::log1p because some compilers
-                             // error out otherwise (E.g. gcc 4.9 in cygwin)
-        if (x <= -37)
-            return exp(x);
-        else if (-37 < x && x <= 18)
-            return log1p(exp(x));
-        else if (18 < x && x <= 33.3)
-            return x + exp(-x);
-        else
-            return x;
-    }
-    
-// ----------------------------------------------------------------------------------------
-
     // Tell us if T is one of the special layer types (i.e. add_layer, repeat, add_tag_layer, or
     // add_skip_layer).
     template <typename T> struct is_nonloss_layer_type : std::false_type {};
@@ -442,24 +425,6 @@ namespace dlib
     {
         return item;
     }
-// ----------------------------------------------------------------------------------------
-
-    inline void randomize_parameters (
-        tensor& params,
-        unsigned long num_inputs_and_outputs,
-        dlib::rand& rnd
-    )
-    {
-        for (auto& val : params)
-        {
-            // Draw a random number to initialize the layer according to formula (16)
-            // from Understanding the difficulty of training deep feedforward neural
-            // networks by Xavier Glorot and Yoshua Bengio.
-            val = 2*rnd.get_random_float()-1;
-            val *= std::sqrt(6.0/(num_inputs_and_outputs));
-        }
-    }
-
 // ----------------------------------------------------------------------------------------
 
     template <typename T>
