+++ dlib/dnn/core_abstract.h
@@ -16,23 +16,6 @@ namespace dlib
 
 // ----------------------------------------------------------------------------------------
 
-    void randomize_parameters (
-        tensor& params,
-        unsigned long num_inputs_and_outputs,
-        dlib::rand& rnd
-    );
-    /*!
-        ensures
-            - This function assigns random values into params based on the given random
-              number generator.  In particular, it uses the parameter initialization method
-              of formula 16 from the paper "Understanding the difficulty of training deep
-              feedforward neural networks" by Xavier Glorot and Yoshua Bengio.
-            - It is assumed that the total number of inputs and outputs from the layer is
-              num_inputs_and_outputs.  That is, you should set num_inputs_and_outputs to
-              the sum of the dimensionalities of the vectors going into and out of the
-              layer that uses params as its parameters.
-    !*/
-
     template <
         typename... T 
         >
@@ -56,15 +39,6 @@ namespace dlib
             - returns a copy of the first thing in the tuple that isn't a std::tuple.
               Essentially, this function calls std::get<0>() recursively on item until
               a non-std::tuple object is found.
-    !*/
-
-    double log1pexp(
-        double x
-    );
-    /*!
-        ensures
-            - returns log(1+exp(x))
-              (except computes it using a numerically accurate method)
     !*/
 
 // ----------------------------------------------------------------------------------------
